+++ lib/taurus/qt/qtdesigner/taurusdesigner.py
@@ -25,8 +25,8 @@
 
 from builtins import str
 import sys
+import click
 import os.path
-import optparse
 
 import taurus
 from taurus.external.qt import Qt
@@ -80,6 +80,8 @@ def get_qtdesigner_bin():
         designer_bin = os.path.join(
             designer_bin, "Designer.app", "Contents", "MacOS")
     elif plat in ("win32", "nt"):
+        # TODO: refactor this to make it Qt-binding agnostic
+        # TODO: check if this works for conda installations on windows
         import PyQt4
         designer_bin = os.path.abspath(os.path.dirname(PyQt4.__file__))
 
@@ -151,31 +153,17 @@ def qtdesigner_start(args, env=None):
     return designer.exitCode()
 
 
-def main(env=None):
-    if env is not None:
-        taurus.info('ignoring obsolete env parameter to qtdesigner_start')
+@click.command('designer')
+@click.argument('ui_files', nargs=-1)
+@click.option("--taurus-path", "tauruspath",
+              metavar='TAURUSPATH',
+              default=None,
+              help="additional directories to look for taurus widgets")
+def designer_cmd(ui_files, tauruspath):
+    """Launch a Taurus-customized Qt Designer application"""
 
-    version = "taurusdesigner %s" % (taurus.Release.version)
-    usage = "Usage: %prog [options] <ui file(s)>"
-    description = "The Qt designer application customized for taurus"
-    parser = optparse.OptionParser(
-        version=version, usage=usage, description=description)
-    parser.add_option("--taurus-path", dest="tauruspath", default="",
-                      help="additional directories to look for taurus widgets")
-    parser.add_option("--qt-designer-path", dest="pyqtdesignerpath", default="",
-                      help="additional directories to look for python qt widgets")
-
-    options, args = parser.parse_args()
-
-    taurus_extra_path = None
-    # Set TAURUSQTDESIGNERPATH
-    if len(options.tauruspath) > 0:
-        taurus_extra_path = options.tauruspath
-
-    if env is None:
-        env = get_taurus_designer_env(taurus_extra_path=taurus_extra_path)
-
-    sys.exit(qtdesigner_start(args, env=env))
+    env = get_taurus_designer_env(taurus_extra_path=tauruspath)
+    sys.exit(qtdesigner_start(ui_files, env=env))
 
 if __name__ == "__main__":
     main()