+++ lib/taurus/qt/qtgui/extra_guiqwt/image.py
@@ -601,7 +601,7 @@ def test1():
     from guiqwt.plot import ImageDialog
     from taurus.qt.qtgui.extra_guiqwt.builder import make
     from taurus.qt.qtgui.application import TaurusApplication
-    app = TaurusApplication()
+    app = TaurusApplication(cmd_line_parser=None)
 
     # define a taurus image
     #model1 = 'sys/tg_test/1/short_image_ro'