+++ lib/taurus/qt/qtgui/input/choicedlg.py
@@ -249,7 +249,7 @@ def testWidget():
     import sys
     from taurus.qt.qtgui.application import TaurusApplication
 
-    app = TaurusApplication(sys.argv)
+    app = TaurusApplication(sys.argv, cmd_line_parser=None)
     w = GraphicalChoiceWidget(None, True)
     w.show()
     sys.exit(app.exec_())
@@ -260,7 +260,7 @@ def main():
     from taurus.qt.qtgui.icon import getCachedPixmap
     from taurus.qt.qtgui.application import TaurusApplication
 
-    app = TaurusApplication(sys.argv)
+    app = TaurusApplication(sys.argv, cmd_line_parser=None)
 
     pixmaps = {}
     choices = [['TaurusForm', 'TaurusTrend'], ['TaurusPlot', 'Qub']]